backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
pydantic[email-validator]==2.5.0
python-Levenshtein==0.23.0
bcrypt==4.0.1
